Added files for Steel OS 3 Build 1816+ (post-1816)
Added lock screen files for post-1816 (otherwise known as Build 1817-????) builds. A new background is to be added, as well. Fun fact: I was originally going to implement the boot screen to be integrated into the lock screen, which is why there is a boot.png file, featuring a static background and copyright info.