Added new background files for post-1816 builds
Added Steel OS 3 Beta 1 Build 1817/later post-3.0.0.1816 builds to the repository. This includes a new background, the *.pdn (which is effectively the Paint.NET equivalent of a PSD file) file used for this background, and a prototype of this background that wasn't used. This background is supposed to be a demo for Steel OS 3's new Aero visual theme (inspired by Windows Vista/7 Aero, but retains Steel OS 2.3's color set), but once again, I am not the best at finding a background for demos like this.